Refactored auth/views.py Refactored bl Moved BL from User model to UserBL
