+++ project/alembic/versions/3a61ff038f2_add_entities.py
@@ -15,7 +15,7 @@ depends_on = None
 from alembic import op
 import sqlalchemy as sa
 import project
-from project.auth.models import User
+from project.models import User
 
 
 def upgrade():
