+++ project/auth/views.py
@@ -1,5 +1,4 @@
-from flask import render_template, Blueprint, flash, request, session, \
-    redirect, url_for
+from flask import render_template, Blueprint, flash, session, redirect, url_for
 from .forms import LoginForm, RegisterForm
 from .decorators import login_required
 from project.models import User
@@ -9,41 +8,42 @@ auth = Blueprint('auth', __name__)
 
 @auth.route('/login', methods=['GET', 'POST'])
 def login():
-    if request.method == 'POST':
-        form = LoginForm(request.form)
-        if form.validate_on_submit():
-            user = User.authenticate(**form.data)
-            if user:
-                session['user_id'] = user.id
-                return redirect(url_for('admin.vacancy_list'))
-            else:
-                flash("Неправильный логин и/или пароль")
-    else:
-        if session.get('user_id'):
+    form = LoginForm()
+    if form.validate_on_submit():
+        user = User.bl.authenticate(**form.data)
+        if user:
+            session['user_id'] = user.id
             return redirect(url_for('admin.vacancy_list'))
-        form = LoginForm()
-    return render_template('login.html',
-                           title='Sign in',
-                           form=form)
+        else:
+            flash("Неправильный логин и/или пароль")
+
+    if session.get('user_id'):
+        return redirect(url_for('admin.vacancy_list'))
+
+    return render_template(
+        'login.html',
+        title='Sign in',
+        form=form,
+    )
 
 
 @auth.route('/register', methods=['GET', 'POST'])
 @login_required
 def register():
-    if request.method == 'POST':
-        form = RegisterForm(request.form)
-        if form.validate_on_submit():
-            userdata = form.data
-            userdata.pop('confirmation')
-            u = User(**userdata)
-            u.save()
-            flash("Вы успешно зарегистрировались")
-            return redirect(url_for('auth.login'))
-    else:
-        form = RegisterForm()
-    return render_template('login.html',
-                           title='Registration',
-                           form=form)
+    form = RegisterForm()
+
+    if form.validate_on_submit():
+        userdata = form.data
+        userdata.pop('confirmation')
+        User.bl.create_user(userdata)
+        flash("Вы успешно зарегистрировались")
+        return redirect(url_for('auth.login'))
+
+    return render_template(
+        'login.html',
+        title='Registration',
+        form=form,
+    )
 
 
 @auth.route('/logout', methods=['GET'])
