+++ project/models.py
@@ -3,8 +3,6 @@ from enum import IntEnum
 from sqlalchemy import Text, ForeignKey, Boolean
 from sqlalchemy.orm import relationship, backref
 from sqlalchemy import Column, Integer, String
-from werkzeug.security import generate_password_hash
-
 from project.bl.utils import Resource
 from project.database import Base, db_session, engine
 from project.lib.orm.types import TypeEnum
@@ -74,14 +72,6 @@ class User(Base):
 
     bl = Resource('bl.user')
 
-    def __init__(self, login, password, email, name=None,
-                 surname=None):
-        self.email = email
-        self.login = login
-        self.name = name
-        self.surname = surname
-        self.set_password(password)
-
     def __repr__(self):
         return '<User {}>'.format(self.get_full_name())
 
@@ -94,9 +84,6 @@ class User(Base):
 
     def is_superuser(self):
         return self.role == self.ROLE.superuser
-
-    def set_password(self, password):
-        self.password = generate_password_hash(password)
 
 
 class City(Base):
